PMM-7259 Move IntegratedAlertingTabs
